+++ NEWS.md
@@ -52,6 +52,8 @@
 
 # pkgbuild 1.0.1
 
+* Preliminary support for rtools 4.0 (#40)
+
 * `compile_dll()` now does not supply compiler flags if there is an existing
   user defined Makevars file.